requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
aiohttp==3.9.5
pydantic==2.9.2
